Create ActiveDirectory4WorkGroups.ps1
If you work in a Workgrouped environment and need to manage local users and groups then this script is for you.

It creates a GUI driven menu to enable you to centrally manage (local) users and groups on a single (or multiple Windows hosts) at the click of one button. 

There must be a user account that is common to all of the hosts when doing multiple hosts, and the password of that account must be the same or you will encounter authentication errors. (Usually this is the localhost\administrator account)

You can supply a CSV file for the script to apply the user/group changes to as many remote servers listed in the csv file as you like.
This does require WinRM to allow the execution of powershell scripts/commands in a pssession to work.